Split Menu button and Menu panel
