+++ src/panel/Menu.jsx
@@ -2,11 +2,9 @@
 import React, { Fragment, useEffect, useState, useRef, useContext } from 'react';
 import ReactDOM from 'react-dom';
 import cx from 'classnames';
-import MenuItem from './menu/MenuItem';
+import MenuPanel from './menu/MenuPanel';
 import Telemetry from 'src/libs/telemetry';
-import { CloseButton, Flex } from 'src/components/ui';
-import { Heart, IconLightbulb, IconEdit, IconMenu } from 'src/components/ui/icons';
-import { PINK_DARK, ACTION_BLUE_BASE } from 'src/libs/colors';
+import { IconMenu } from 'src/components/ui/icons';
 import { DeviceContext } from 'src/libs/device';
 
 const Menu = () => {
@@ -33,11 +31,6 @@ const Menu = () => {
     setIsOpen(false);
   };
 
-  const navTo = (url, options) => {
-    close();
-    window.app.navigateTo(url, options);
-  };
-
   return (
     <Fragment>
       <button
@@ -49,49 +42,7 @@ const Menu = () => {
         {isMobile ? <IconMenu /> : <IconMenu width={16} height={16} />}
       </button>
 
-      {isOpen &&
-        ReactDOM.createPortal(
-          <div className="menu">
-            <div className="menu__overlay" onClick={close} />
-
-            <div className="menu__panel">
-              <Flex className="menu-top u-mb-l">
-                <CloseButton circle onClick={close} />
-              </Flex>
-              <div className="menu-items">
-                <MenuItem
-                  onClick={e => {
-                    e.preventDefault();
-                    Telemetry.add(Telemetry.MENU_FAVORITE);
-                    navTo('/favs/');
-                  }}
-                  icon={<Heart width={16} color={PINK_DARK} />}
-                >
-                  {_('My favorites', 'menu')}
-                </MenuItem>
-                <MenuItem
-                  href="https://about.qwant.com/legal/terms-of-service/qwant-maps/"
-                  outsideLink
-                  icon={<IconLightbulb width={16} fill={ACTION_BLUE_BASE} />}
-                >
-                  <span
-                    dangerouslySetInnerHTML={{
-                      __html: _('Terms of service Qwant&nbsp;Maps', 'menu'),
-                    }}
-                  />
-                </MenuItem>
-                <MenuItem
-                  href="https://github.com/Qwant/qwantmaps/blob/master/contributing.md"
-                  outsideLink
-                  icon={<IconEdit width={16} fill={ACTION_BLUE_BASE} />}
-                >
-                  {_('How to contribute', 'menu')}
-                </MenuItem>
-              </div>
-            </div>
-          </div>,
-          menuContainer.current
-        )}
+      {isOpen && ReactDOM.createPortal(<MenuPanel close={close} />, menuContainer.current)}
     </Fragment>
   );
 };
